Refactored, updated images
