+++ src/App.js
@@ -1,12 +1,16 @@
-import { GameDetails } from "./components/GameDetails";
+import { Intro } from "./components/Intro";
 import { IllustrationWrapper } from "./components/Illustration";
 import { Layout } from "./components/Layout";
 import "./styles/index.css";
+import HangmanGame from "./components/HangmanGame";
 
 const App = () => (
   <Layout>
-    <IllustrationWrapper lives={2} />
-    <GameDetails />
+    <IllustrationWrapper livesLeft={6} />
+    <div className="rules__wrapper">
+      <Intro />
+      <HangmanGame />
+    </div>
   </Layout>
 );
