+++ src/styles/index.css
@@ -23,21 +23,20 @@ body {
 .illustration__wrapper {
   display: flex;
   flex-grow: 1;
-  height: 100%;
+  min-width: 200px;
   max-width: 600px;
   position: relative;
+  overflow: hidden;
 }
 
 .rules__wrapper {
   display: flex;
   flex-direction: column;
   align-items: flex-start;
-
-  height: 100%;
   gap: 10px;
   flex-grow: 2;
   max-width: 800px;
-  padding: 100px 0;
+  padding: 100px 40px 50px 0;
 }
 
 .word__wrapper {
@@ -140,4 +139,10 @@ body {
   border-color: #f0f0f0;
   background-color: #f0f0f0;
   transition: all 0.3s ease;
+}
+
+.guessedletters {
+  font-size: 42px;
+  margin-top: 0;
+  font-family: monospace;
 }
